+++ changelog.md
@@ -2,6 +2,8 @@
 
 - Minor code refactoring
 - Added the ability to get a map of registered commands
+- Added command description annotation
+- Added methods for getting values from command annotations
 - Fixed plugin example
 - Fixed copying configs from the plugin